requirements.txt
textual>=1.0.0
websockets>=12.0

httpx>=0.27.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.0